Move to TreeError.tsx
