+++ packages/kiwi-react/src/bricks/Tree.tsx
@@ -6,9 +6,8 @@ import cx from "classnames";
 import { Role } from "@ariakit/react/role";
 import { useCompositeStore, Composite } from "@ariakit/react/composite";
 import { forwardRef, type BaseProps } from "./~utils.js";
+import { Root as TreeError } from "./TreeError.js";
 import { Root as TreeItemRoot, Action as TreeItemAction } from "./TreeItem.js";
-import { StatusWarning } from "./Icon.js";
-import { Text } from "./Text.js";
 
 // ----------------------------------------------------------------------------
 
@@ -46,33 +45,6 @@ const Tree = forwardRef<"div", TreeProps>((props, forwardedRef) => {
 	);
 });
 DEV: Tree.displayName = "Tree.Root";
-
-// ----------------------------------------------------------------------------
-
-interface TreeErrorProps extends BaseProps {
-	/**
-	 * Label for the tree header indicating the number of errors displayed.
-	 */
-	label?: React.ReactNode;
-}
-
-const TreeError = forwardRef<"div", TreeErrorProps>((props, forwardedRef) => {
-	const { label, ...rest } = props;
-	return (
-		<Role.div
-			{...rest}
-			className={cx("🥝-tree-error", props.className)}
-			ref={forwardedRef}
-		>
-			<div className="🥝-tree-error-header">
-				<StatusWarning />
-				<Text variant="body-sm">{label}</Text>
-			</div>
-			<div className="🥝-tree-error-items">{props.children}</div>
-		</Role.div>
-	);
-});
-DEV: TreeError.displayName = "Tree.Error";
 
 // ----------------------------------------------------------------------------
