+++ packages/kiwi-react/src/bricks/styles.css
@@ -32,6 +32,7 @@
 @import "./TextBox.css" layer(itwinui.components);
 @import "./Tooltip.css" layer(itwinui.components);
 @import "./Tree.css" layer(itwinui.components);
+@import "./TreeError.css" layer(itwinui.components);
 @import "./TreeItem.css" layer(itwinui.components);
 
 @import "./Field.css" layer(itwinui.components);
